fix: complete Kotlin coroutine review fixes
