+++ gluecodium/src/main/resources/templates/kotlin/KotlinClass.mustache
@@ -126,6 +126,10 @@
 {{/if}}
 {{/properties}}
 {{/ifPredicate}}
+{{#if staticCoroutineMembers}}
+
+{{{staticCoroutineMembers}}}
+{{/if}}
     }
 {{/ifPredicate}}
 {{#if coroutineMembers}}
