+++ gluecodium/src/main/resources/templates/kotlin/KotlinStruct.mustache
@@ -199,6 +199,11 @@
 {{/unless}}
 {{/functions}}
 
+{{#if coroutineMembers}}
+
+{{{coroutineMembers}}}
+{{/if}}
+
 {{#eval "optimizedLists" path}}{{#this}}
 {{#elementType}}{{prefixPartial "kotlin/KotlinLazyNativeList" "    "}}{{/elementType}}
 {{/this}}{{/eval}}
@@ -221,6 +226,10 @@
 {{#unless external.kotlin.name}}
 {{#if attributes.serializable}}{{prefixPartial "kotlin/KotlinParcelableCreator" "        "}}
 {{/if}}{{/unless}}
+{{#if staticCoroutineMembers}}
+
+{{{staticCoroutineMembers}}}
+{{/if}}
     }
 {{/ifPredicate}}
 }
