+++ gluecodium/src/test/resources/smoke/kotlin_coroutines/output/android-kotlin/com/example/coroutines/KotlinCoroutines.kt
@@ -1,12 +1,94 @@
 // Generated by Gluecodium. Coroutine (suspend) support for @KotlinCoroutine functions.
 
-@file:JvmName("CoroutineClientKotlinCoroutines")
+@file:JvmName("CoroutineClientRemoteServiceKotlinCoroutines")
 
 package com.example.coroutines
 
 import kotlin.coroutines.resume
 import kotlin.coroutines.resumeWithException
 import kotlinx.coroutines.suspendCancellableCoroutine
+
+/** Thrown when the underlying SDK callback breaks its (error XOR result) contract. */
+internal class RemoteServiceSdkContractViolationException(message: String) : IllegalStateException(message)
+
+/**
+ * Generic bridge that adapts a callback-style [startOperation] into a `suspend` call returning [Result].
+ *
+ * Behavior:
+ * - resumes with [Result.success] carrying the callback value,
+ * - resumes with [Result.failure] produced by [mapDomainError] on a domain error,
+ * - fails with [RemoteServiceSdkContractViolationException] if the callback breaks the (error XOR result) contract,
+ * - propagates a synchronous throwable from [startOperation],
+ * - cancels the underlying task via [cancelOperation] when the coroutine is cancelled before the
+ *   callback resolves. The cancellation path shares the `resolved` guard with the callback, so a
+ *   cancellation racing a completing callback never cancels an already-finished task.
+ */
+internal suspend fun <E, T, H> remoteServiceAwaitResultBridge(
+    startOperation: (callback: (E?, T?) -> Unit) -> H,
+    mapDomainError: (E) -> Throwable,
+    cancelOperation: (H) -> Unit,
+): Result<T> =
+    suspendCancellableCoroutine { continuation ->
+        val resolved = java.util.concurrent.atomic.AtomicBoolean(false)
+        val handle: H? =
+            try {
+                startOperation { error, value ->
+                    if (resolved.compareAndSet(false, true)) {
+                        continuation.resume(
+                            when {
+                                error == null && value != null -> Result.success(value)
+                                error != null && value == null -> Result.failure(mapDomainError(error))
+                                error == null && value == null ->
+                                    Result.failure(RemoteServiceSdkContractViolationException("SDK contract violation: both error and value are null"))
+                                else ->
+                                    Result.failure(RemoteServiceSdkContractViolationException("SDK contract violation: both error and value are non-null"))
+                            }
+                        )
+                    }
+                }
+            } catch (throwable: Throwable) {
+                if (resolved.compareAndSet(false, true)) {
+                    continuation.resume(Result.failure(throwable))
+                }
+                null
+            }
+        if (handle != null) {
+            continuation.invokeOnCancellation {
+                if (resolved.compareAndSet(false, true)) {
+                    cancelOperation(handle)
+                }
+            }
+        }
+    }
+
+/** Adapts a callback without a domain-error channel into a direct suspending value. */
+internal suspend fun <T, H> remoteServiceAwaitValueBridge(
+    startOperation: (callback: (T) -> Unit) -> H,
+    cancelOperation: (H) -> Unit,
+): T =
+    suspendCancellableCoroutine { continuation ->
+        val resolved = java.util.concurrent.atomic.AtomicBoolean(false)
+        val handle: H? =
+            try {
+                startOperation { value ->
+                    if (resolved.compareAndSet(false, true)) {
+                        continuation.resume(value)
+                    }
+                }
+            } catch (throwable: Throwable) {
+                if (resolved.compareAndSet(false, true)) {
+                    continuation.resumeWithException(throwable)
+                }
+                null
+            }
+        if (handle != null) {
+            continuation.invokeOnCancellation {
+                if (resolved.compareAndSet(false, true)) {
+                    cancelOperation(handle)
+                }
+            }
+        }
+    }
 
 /** Thrown when the underlying SDK callback breaks its (error XOR result) contract. */
 internal class CoroutineClientSdkContractViolationException(message: String) : IllegalStateException(message)
