+++ gluecodium/src/test/resources/smoke/kotlin_coroutines/output/android-kotlin/com/example/coroutines/RemoteService.kt
@@ -1,0 +1,66 @@
+/*
+
+ *
+ */
+
+@file:JvmName("RemoteServiceExtensions")
+
+
+package com.example.coroutines
+
+import com.example.NativeBase
+
+fun interface RemoteService {
+
+    /**
+     * Checks whether the remote service is currently reachable.
+     * @param callback
+     */
+
+    fun checkConnection(callback: ErrorOnlyCallback) : Unit
+
+
+
+    /** Typed exception wrapper preserving the callback error value. */
+    public class CheckConnectionException(public val error: OperationError) : Exception(error.toString())
+
+    /**
+     * Coroutine (`suspend`) variant of `checkConnection`.
+     *
+     * Suspends the calling coroutine until the operation completes instead of taking a callback.
+     *
+     * Checks whether the remote service is currently reachable.
+     * @return [Result] holding when the operation completes on success, or a failure carrying CheckConnectionException on error.
+     */
+    public suspend fun checkConnection(): Result<Unit> =
+            remoteServiceAwaitResultBridge(
+            startOperation = { callback -> this.checkConnection(ErrorOnlyCallback { error -> callback(error, if (error == null) Unit else null) }) },
+            mapDomainError = { error -> CheckConnectionException(error) },
+            cancelOperation = { handle -> Unit },
+        )
+
+}
+
+/**
+ * @suppress
+ *
+ * This class is used to represent C++ implementations of the interface or lambda in Kotlin.
+ * It is instantiated by JNI and should not be used by the end users.
+ */
+private class RemoteServiceImpl : NativeBase, RemoteService {
+    protected constructor(nativeHandle: Long, @Suppress("UNUSED_PARAMETER") tag: Any?)
+        : super(nativeHandle, { disposeNativeHandle(it) }) {}
+
+    /**
+     * Checks whether the remote service is currently reachable.
+     * @param callback
+     */
+
+    override external fun checkConnection(callback: ErrorOnlyCallback) : Unit
+
+
+
+    companion object {
+        @JvmStatic private external fun disposeNativeHandle(nativeHandle: Long)
+    }
+}
